+++ app/src/main/res/layout/fragment_cosa_prepariamo_oggi.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.cosa_preparare_oggi.CosaPreparareOggiFragment">
+    tools:context=".ui.cosa_prepariamo_oggi.CosaPrepariamoOggiFragment">
 
 
     <ImageButton
